requirements.txt
tensorflow-cpu==2.18.0
keras==3.10.0
streamlit
pillow
numpy